+++ SourceCode/Data/MeetingParticipant.cs
@@ -57,6 +57,8 @@ public static class MeetingParticipantMapping
 public static class MeetingParticipantExtensions
 {
     public static bool IsCancelled(this MeetingParticipant participant) => participant.CancellationTime.HasValue;
+
+    public static bool IsParticipating(this MeetingParticipant participant) => !participant.IsCancelled();
     public static DateTime FirstParticpationDate(this MeetingParticipant participant) =>
         participant.ParticipateDay1 ? participant.Meeting.StartDate :
         participant.ParticipateDay2 ? participant.Meeting.StartDate.AddDays(1) :